finalize selectAll.ts, import into App.tsx, and clean up updateOption function
